requirements.txt
chromadb
pypdf
python-dotenv
sentence-transformers
google-genai
torch
streamlit